refactor: refactor backend FastAPI project into modular structure
