+++ backend/src/news/models.py
@@ -1,0 +1,39 @@
+from sqlalchemy import Column, ForeignKey, Integer, String, Table, Text
+from sqlalchemy.orm import relationship
+
+from src.database import Base
+
+user_news_association_table = Table(
+    "user_news_upvotes",
+    Base.metadata,
+    Column(
+        "user_id",
+        Integer,
+        ForeignKey("users.id"),
+        primary_key=True,
+    ),
+    Column(
+        "news_articles_id",
+        Integer,
+        ForeignKey("news_articles.id"),
+        primary_key=True,
+    ),
+)
+
+
+class NewsArticle(Base):
+    __tablename__ = "news_articles"
+
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    url = Column(String, unique=True, nullable=False)
+    title = Column(String, nullable=False)
+    time = Column(String, nullable=False)
+    content = Column(Text, nullable=False)
+    summary = Column(Text, nullable=False)
+    reason = Column(Text, nullable=False)
+
+    upvoted_by_users = relationship(
+        "User",
+        secondary=user_news_association_table,
+        back_populates="upvoted_news",
+    )
