+++ backend/src/news/schemas.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+
+class PromptRequest(BaseModel):
+    prompt: str
+
+
+class NewsSummaryRequestSchema(BaseModel):
+    content: str
